fix(commands): add top-level compile command
