+++ src/types/command/commandFactory.ts
@@ -3,7 +3,6 @@ import { AddTargetCommand } from '../../commands/add/addTargetCommand'
 import { BuildCommand } from '../../commands/build/buildCommand'
 import { CompileBuildCommand } from '../../commands/build/compileBuildCommand'
 import { CompileCommand } from '../../commands/compile/compileCommand'
-import { CompileSingleFileCommand } from '../../commands/compile/compileSingleFileCommand'
 import { CreateContextCommand } from '../../commands/context/createContextCommand'
 import { DeleteContextCommand } from '../../commands/context/deleteContextCommand'
 import { EditContextCommand } from '../../commands/context/editContextCommand'
@@ -29,9 +28,6 @@ export const commandFactory = new Map<string, Function>([
   ['add cred', (args: string[]) => new AddCredentialCommand(args)],
   ['build', (args: string[]) => new BuildCommand(args)],
   ['compile', (args: string[]) => new CompileCommand(args)],
-  ['compile job', (args: string[]) => new CompileSingleFileCommand(args)],
-  ['compile service', (args: string[]) => new CompileSingleFileCommand(args)],
-  ['compile identify', (args: string[]) => new CompileSingleFileCommand(args)],
   ['compilebuild', (args: string[]) => new CompileBuildCommand(args)],
   [
     'compilebuilddeploy',
